Create seed/index.ts
